Moved stats relevant function to Statistics module
